refactoring
- fixed bug webview not working anymore
- move most part of connection directory into settings
