+++ src/extension/connection/index.ts
@@ -1,5 +1,3 @@
-export * from "./src/commands";
-
 export const enum ConnectionCommands {
     CREATE   = "VSQlik.Connection.Create",
     SETTINGS = "VSQlik.Connection.Settings",
