--- src/extension/connection/src/settings/index.ts
@@ -1,1 +1,0 @@
-export * from "./src/connection.view";
