--- src/extension/connection/src/utils/index.ts
@@ -1,1 +1,0 @@
-export * from "./src/connection.service";
