+++ src/extension/settings/utils/index.ts
@@ -1,2 +1,3 @@
 export * from "./src/connection-create";
 export * from "./src/connection-settings";
+export * from "./src/settings.service";
