+++ src/extension/settings/utils/src/connection-settings.ts
@@ -1,4 +1,4 @@
-import { ConnectionWebview } from "../../settings";
+import { ConnectionWebview } from "../../webview/src/connection.view";
 
 export function ConnectionSettingsCommands() {
     const view = new ConnectionWebview();
