+++ src/extension/settings/webview/src/connection.html
